+++ src/solver/shared.c
@@ -4,7 +4,7 @@
 solver_t *solver;
 size_t gdata_size;
 
-uint32_t *gdata_get_hash(gdata_t *gd)
+uint64_t *gdata_get_hash(gdata_t *gd)
 {
 	if(solver->uses_zobrist)
 		return &(gd->hash);
@@ -17,7 +17,7 @@ bool make_new_move(gdata_t *child, gdata_t *gd, int move)
 
 	memcpy(child, gd, gdata_size);
 
-	uint32_t *hp = gdata_get_hash(child);
+	uint64_t *hp = gdata_get_hash(child);
 	solver->make_move(&(child->pos), move, hp);
 	child->move_index = move;
