+++ src/solver/shared.h
@@ -18,14 +18,14 @@ extern size_t gdata_size;
 
 typedef struct
 {
-	uint32_t hash;
+	uint64_t hash;
 	float score;
 	int move_index;
 	//uint8_t move_ct;
 	uint8_t pos[];
 } gdata_t;
 
-uint32_t *gdata_get_hash(gdata_t *gd);
+uint64_t *gdata_get_hash(gdata_t *gd);
 bool make_new_move(gdata_t *child, gdata_t *gd, int move);
 
 #endif	//SHARED_H_
